Coordinates to double

+++ ProjectTemplate_v2/Models/HumiditySensor.cs
@@ -11,7 +11,7 @@ namespace ProjectTemplate_v2
         private decimal minValue;
         private decimal maxValue;
 
-        public HumiditySensor(string name, string url, string description, decimal latitude, decimal longitude,bool followed, decimal minValue, decimal maxValue)
+        public HumiditySensor(string name, string url, string description, double latitude, double longitude,bool followed, decimal minValue, decimal maxValue)
             : base(name, url, description, latitude, longitude,followed)
         {
             MinValue = minValue;
